move codebase to src directory
