--- info.h
@@ -1,5 +1,0 @@
-#pragma once
-#include <vector>
-#include <string>
-
-std::vector<std::string> getInfo();
