+++ src/info.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <vector>
+#include <string>
+
+std::vector<std::string> getInfo();
