+++ src/linux/linux_info.h
@@ -1,0 +1,21 @@
+#pragma once
+#include <string>
+
+struct Memory
+{
+    unsigned long total;
+    unsigned long used;
+};
+
+class Linux
+{
+private:
+public:
+    Memory memory();
+    std::string kernel();
+    std::string os();
+    std::string hostname();
+    std::string username();
+    std::string shell();
+    unsigned long long uptime();
+};
